requirements.txt
streamlit
nltk
pandas
python-dotenv
google-generativeai